Add files via upload
The main script is a shell script consistes of all steps: 
0) Building database
1) Alignment step
2) Collecting reads
3) Assembly using SPAdes
4) Final editing using R

We run each script step by step, where in some steps we can use array jobs. It is mentioned in the scripts, where we can use  array jobs or not. Please note that you have to comment in and comment out each steps when you are excuting different steps.